requirements.txt
pandas
numpy
torch
matplotlib
Pillow
torchvision
scikit-learn